Project/requirements.txt
matplotlib~=3.10.1
numpy~=2.2.4
pandas~=2.2.3
seaborn~=0.13.2
scikit-learn~=1.6.1
scipy~=1.15.2
imbalanced-learn~=0.13.0
xgboost~=3.0.0
lightgbm~=4.6.0
Flask~=3.1.0
streamlit~=1.44.1
requests~=2.32.3
pillow~=11.1.0
